# [FileFormats.MPS] improve performance of MPS writer