@odow 2024-02-04
Now to
```julia
julia> GC.gc(); @time main(100_000);
  1.362330 seconds (10.25 M allocations: 526.213 MiB, 9.37% gc time)
```