@odow 2024-02-04
Now to
```julia
julia> GC.gc(); @time main(100_000);
  1.184845 seconds (8.85 M allocations: 475.188 MiB, 11.24% gc time)
```