@odow 2024-02-04
now to
```
julia> GC.gc(); @time main(100_000);
  1.199502 seconds (7.50 M allocations: 448.509 MiB, 10.95% gc time)
```