@odow 2024-02-04
Now to
```julia
julia> GC.gc(); @time main(100_000);
  1.116765 seconds (6.90 M allocations: 436.318 MiB, 11.98% gc time)
```