@odow 2024-02-05
Apart from the unique name stuff, the body of `write` is now completely type-stable (no red bars):

![image](https://github.com/jump-dev/MathOptInterface.jl/assets/8177701/ec0abdb8-9821-426f-9869-81e7487e9023)

Almost all the time now is turning `Int` and `Float64` to string, variable name lookups, and I/O locks.