@odow 2024-02-05
Miles' example is now:
```
(base) oscar@Oscars-MBP /tmp % julia --project=opt ~/Downloads/generate-flow-oscar.jl --output_file mcflow.mps.gz --num_commodities 1000 --num_warehouses 20 --num_stores 200 --seed 1
Parsed args:
  output_file  =>  mcflow.mps.gz
  num_warehouses  =>  20
  num_factories_per_commodity  =>  5
  seed  =>  1
  additional_overtime_cost  =>  0.3
  num_stores  =>  200
  num_commodities  =>  1000
building model ...
model built
 56.593121 seconds (141.45 M allocations: 11.922 GiB, 20.53% gc time, 13.50% compilation time)
```

Original was
```
(base) oscar@Oscars-MBP /tmp % julia --project=moi-release ~/Downloads/generate-flow-oscar.jl --output_file mcflow.mps.gz --num_commodities 1000 --num_warehouses 20 --num_stores 200 --seed 1
Parsed args:
  output_file  =>  mcflow.mps.gz
  num_warehouses  =>  20
  num_factories_per_commodity  =>  5
  seed  =>  1
  additional_overtime_cost  =>  0.3
  num_stores  =>  200
  num_commodities  =>  1000
building model ...
model built
 68.109835 seconds (383.89 M allocations: 19.984 GiB, 20.36% gc time, 6.59% compilation time)
```

So nearly 1/3 the allocations and 60% of the total allocated.

I've been mucking around with storing the indices instead of names, but yet to find something that makes a difference.