@odow 2024-02-05
I tried replacing the `row_name` with an index, but it didn't help things. Only one copy of the name is stored in `coefficients` because the strings must be stored by reference.