@odow 2024-02-06
To do better we'd need a much larger rewrite that stored the problem in
column order to start with instead of the current constraint based
approach.  You probably still have a 2X memory penalty as we transpose the
matrix.

On Tue, 6 Feb 2024, 2:01 pm Miles Lubin, ***@***.***> wrote:

> Nice improvements!
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/jump-dev/MathOptInterface.jl/pull/2418#issuecomment-1928586707>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AB6MQJMRAEIF35ZII3C65XTYSF6GFAVCNFSM6AAAAABCZFSEKSVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTSMRYGU4DMNZQG4>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>