@odow 2024-02-06
Merging this as an improvement for now. But I have a plan for a larger change.