@blegat 2024-02-07
> To do better we'd need a much larger rewrite that stored the problem in
column order to start with instead of the current constraint based
approach.  You probably still have a 2X memory penalty as we transpose the
matrix.

This sounds like what `Utilities.MatrixOfConstraints` and `Utilities.MutableSparseMatrixCSC` are trying to resolve. See my talk at the 2021 INFORMS Annual Meeting (the slides are on my website).