@odow 2024-02-07
MatrixOfConstraints can't incrementally build a CSC though right? It just does it in a `copy_to`.

The main issue with the current design is that when writing, you can the CSC, and when reading, you want the CSR. At the moment, we use a single data structure, but it'd be nice to switch between them depending on whether you're reading or writing.