@blegat 2024-02-08
No, it's done incrementally. `Utilities.MutableSparseMatrixCSC` receives the CSR form (because you receive constraint by constraint so row by row) and but it's actually building a CSC form.